Project structure and lint config
