+++ src/routine.js
@@ -1,8 +1,10 @@
 const client = require('./client')
 
+
 const { ASANA_PROJECT = '1152701043959235' } = process.env
 
-const pastFilter = task => new Date(`${task.due_at ? task.due_at : `${task.due_on}T23:59:59.999Z`}`) < new Date()
+const pastFilter = (task) => new Date(`${task.due_at ? task.due_at : `${task.due_on}T23:59:59}`}`)
+  < new Date()
 
 const markPastDue = (project_id = ASANA_PROJECT) => client.projects.tasks(project_id, {
   completed_since: 'now',
@@ -14,8 +16,10 @@ const markPastDue = (project_id = ASANA_PROJECT) => client.projects.tasks(projec
 
 if (typeof require !== 'undefined' && require.main === module) {
   markPastDue().then((res) => {
+    /* eslint-disable */
     console.log(res)
     console.log(`Daily sweeping done! ${res.length} past time slots marked as complete!`)
+    /* eslint-enable */
     process.exit(0)
   }).catch((err) => {
     console.error(err)
